requirements.txt
python-dotenv
langgraph
langchain_community
langchain_core
langchain-google-genai
langchain-groq
langchain-text-splitters
tavily-python
uvicorn
pydantic
fastapi
pypdf
chromadb
APScheduler
python-multipart
pillow
google-genai
google-generativeai
